backend/pyproject.toml
[tool.poetry]
name = "revtrust-backend"
version = "0.1.0"
description = "RevTrust backend API"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
anthropic = "^0.8.0"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
pandas = "^2.1.0"
openpyxl = "^3.1.2"
prisma = "^0.11.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
